extract `ImageCommentsUIIntegrationTests` helpers into separate classes
