+++ EssentialApp/EssentialAppTests/Helpers/ImageCommentsUIIntegrationTests+Localization.swift
@@ -1,0 +1,24 @@
+//
+//  ImageCommentsUIIntegrationTests+Localization.swift
+//  EssentialAppTests
+//
+//  Created by Sebastian Vidrea on 09.04.2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import XCTest
+import EssentialFeed
+
+extension ImageCommentsUIIntegrationTests {
+
+	func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
+		let table = "ImageComments"
+		let bundle = Bundle(for: ImageCommentsPresenter.self)
+		let value = bundle.localizedString(forKey: key, value: nil, table: table)
+		if value == key {
+			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
+		}
+		return value
+	}
+
+}
